Refactored the code base  - Moved all functionalities into seperate modules Added styling to the app
