+++ src/index.html
@@ -5,20 +5,18 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" href="style.css">
-  <script src="/node_modules/fuse.js/dist/fuse.js" defer type="module"></script>
-  <script type="module" src="/src/modules/chromeTab.js" defer ></script>
-  <title>My Bookmark Manager chrome extension</title>
+  <script type="module" src="/src/index.js" defer></script>
+  <title>Search Bookmarks</title>
 </head>
 <body>
-  <h1>Hello, and welcome to study pilot</h1>
-  <div>
-    <form autocomplete="off" id="my-form">
-      <label for="input">My favorite fruit:</label>
-      <div>            
-          <input type="text" id="input">
-      </div>
-      <button type="submit">Visit site</button>
-  </form>
+  <div class="container">
+    <div class="inner-container">
+      <h1>Find Your Bookmark</h1>
+      <form autocomplete="off" id="my-form"> 
+        <div><input type="text" id="input" placeholder="Enter name...."></div>
+        <button type="submit" id="go">Visit site</button>
+      </form>
+    </div>
   </div>
 </body>
 </html>
