+++ src/modules/close.js
@@ -1,0 +1,7 @@
+const closeList = () => {
+  const suggestions = document.getElementById('suggestions');
+  if (suggestions) {
+    suggestions.parentNode.removeChild(suggestions);
+  }
+};
+export default closeList;
